+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add VRF BGP IPv4/IPv6 import/export route targets
 - Add service EPG policy
 - Add trust control policy
+- Add redirect backup policy
 
 ## 0.3.3